organize codebase and fix bugs
